Remove appveyor from testing
Very old versions of the site were tested with Appveyor,
a service that could test to make sure build steps ran on Windows.
This isn't really necessary now that the site is Dockerized.

Appveyor hasn't run in over 4 years probably when we moved the
organization from pythonsd -> sandiegopython.
I don't think anybody noticed which is a pretty strong indication
it isn't needed.