Tailwind Upgrade 3.x -> 4.0
Tailwind 4.0 just dropped yesterday so I'm fine waiting on merging this
change to see if any issues shake out more generally.

The tailwind config file is no longer needed and it uses CSS config.
The CLI for tailwind is now a separate package.

I followed the upgrade guide/script for this upgrade:
https://tailwindcss.com/docs/upgrade-guide